Run your postgres with HA using stolon on k8s
This project was born out of a need to be able to run postgres HA in k8s
with a constraint to be able to run some database schema initilisation
scripts/sql to be run as an one time process/effort. Currently stolon
overrides the entrypoint to run its own stolo0n stuff and thus adding
the sql/shell scripts to bootstrap the postgres with schema fails.